Twice the bits, twice the blood? (64-bit build)
Once again, these patches allow the game to build on 64-bit, plus
a shell script to choose which binary to launch. makeit.sh is
included because changing the BINDIR didn't seem to have any effect
within the makefile itself, so I use it to put 32- and 64-bit
object files in different locations. It's a convenience thing,
simply using "make" will still work.