Fixed a few issues
Got rid of an include of postal_limits.h which I forgot to commit,
and which is not necessary.
Switched from str2int to strcmp which is a much uglier construct
but appeases Microsoft's eternal incompetence.
Removed a line of code which enables the invisible final item in
the options menu.