Create new Sprys!
... with bad colours and parts missing. The code to convert a PImage
to an RImage is kinda ugly too. Points for me to work on, and in fact
I will work on those, once this is committed.

I already fixed the Python segfault I mentioned on the mailing list.
Sprys and SWIG both think they own Sprites, and will both try to
de-allocate them. My solution is to convert the Spry back into a
list before the function ends so that the Spry will be emptied and
SWIG can take care of de-allocation.

Oh, and I haven't hooked this up to the GUI yet. You will need to
import psak to a Python2 REPL as a module and use pngToSpry. Or
just wait for me to add it to the GUI.