gcc as CC in makefile; "ameliorate" C library
The makefile now uses gcc to compile C files and saves g++ for
C++ code and linking. I've added the "ameliorate" C library (the
name is completely irrelevant) which I am hoping to use for
PSAK to give it a bit of a speed boost in converting PNG to Spry,
because chopping up the image is proving to be too slow in Python.
I haven't tested the new library yet, though.