Fixed am_chop for real this time
I was getting the data back to Python as a C string. Who's got two
thumbs and forgot that \0 is a string terminator? This guy.
So after all my truly dumb mistakes, that one still made things
extremely difficult. This was a very frustrating issue. But it's
over now.
I'm a bit upset that my memcmp optimisation wouldn't work. I don't
know what the issue is, but I ended up shrinking the box vertically
the same way I was already shrinking it horizontally and even though
that's not the fastest way I could theoretically do it, this
implementation still runs circles around anything I was trying to do
in Python, so I'm just going to leave it like this, at least for now.