Made wkinter optional
I'm not entirely sure if wkinter actually works on Windows.
Just to be on the safe side I put a try..except around the
import. If it raises an exception then psak will skip the
GUI. So you can import it even if you don't have the
required modules for the GUI.