PSAK is dead, long live hoodcraft
This script makes it easier to make a custom level. Just make the
graphics in 24-bit colour space and it will handle all the indexing
and RSPiX formats for you. Save the base image as a BMP and the rest
as PNG files.

You still need to provide collision data. And I have no idea how to
make that yet.